backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
openai==1.3.7
whisper==1.1.10
ffmpeg-python==0.2.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
stripe==7.8.0
boto3==1.34.0
Pillow==10.1.0
